+++ tests/functional_tests/sync_all_tester.hpp
@@ -41,12 +41,12 @@ class SyncAllTester : public Tester {
   virtual ~SyncAllTester();
 
  protected:
-  virtual void resetBuffers(uint64_t size) override;
+  virtual void resetBuffers(size_t size) override;
 
   virtual void launchKernel(dim3 gridSize, dim3 blockSize, int loop,
-                            uint64_t size) override;
+                            size_t size) override;
 
-  virtual void verifyResults(uint64_t size) override;
+  virtual void verifyResults(size_t size) override;
 };
 
 #endif
